+++ Flash-Point-Fire-Rescue/Assets/Scripts/WallsManager.cs
@@ -48,7 +48,7 @@ public class WallsManager : MonoBehaviour
     public void DamageWall(int x, int y, int direction)
     {
         // Posición base igual que usarías para el fuego
-        Vector3 basePosition = new Vector3(x*4, 5, y*4);
+        Vector3 basePosition = new Vector3(x*4, 4.3f, y*4);
 
         // Instanciar un fuego temporal SOLO para obtener su posición real
         GameObject fireTemp = Instantiate(firePrefab2, basePosition, Quaternion.identity);